28/7/2022

+++ crud/delete.php
@@ -1,28 +1,32 @@
 
 
 <?php 
+session_start();
+if(isset($_SESSION["user"]))
+{ 
+        $conn=mysqli_connect('localhost','root','','db'); 
 
-include_once'db.php'; 
+        if (isset($_GET['id'])) 
+        {
 
-if (isset($_GET['id'])) {
+            $id = $_GET['id'];
 
-    $id = $_GET['id'];
+            $sql = "DELETE FROM `crud` WHERE `id`='$id'";
 
-    $sql = "DELETE FROM `crud` WHERE `id`='$id'";
+             $result = $conn->query($sql);
 
-     $result = $conn->query($sql);
+             if ($result == TRUE) {
 
-     if ($result == TRUE) {
+                echo "Record deleted successfully.";
 
-        echo "Record deleted successfully.";
+            }else{
 
-    }else{
+                echo "Error:" . $sql . "<br>" . $conn->error;
 
-        echo "Error:" . $sql . "<br>" . $conn->error;
+            }
+            header('Location:view1.php');
 
-    }
-    header('Location:view1.php');
-
-} 
+        } 
+}
 
 ?>
